net/rds: Restore MR use-once semantics
Commit "rds: new extension header: rdma bytes" has a bug. When
parsing an RDS extension header entry, if the data part of it contains
two consecutive zeros, the parser will incorrectly assume that the two
zeros designate the end of the extension. As consequence, the latter
part of the previous extension header is overwritten.

In particular, if RDMA is used on an MR that is created with the
RDS_RDMA_USE_ONCE flag, the bcopy send telling the requester that the
RDMA has completed (RDMA ACK message or immediate data in ED
terminology), will be sent with the RDS_EXTHDR_RDMA extension
header. This header is four bytes long and contains the rkey of the MR
that was read/written by the RDMA.

After the RDS_EXTHDR_RDMA extension header has been written to the RDS
header, it contains the five bytes:

2, rkey3, rkey2, rkey1, rkey0

2 is the encoding for RDS_EXTHDR_RDMA. Since the RDS wire protocol
encodes multi-byte entities in big endian format, the rkey3 is the
most significant byte (MSB) of the rkey and rkey0 is the least
significant byte of the rkey.

Commit "rds: new extension header: rdma bytes" added support for
another extension header, namely RDS_EXTHDR_RDMA_BYTES (4). This is an
eight-byte extension, with the RDMA size in the first four bytes, then
a byte designating RDMA Read (0x2) or Write (0x1). Thereafter, 3
padding bytes.

If the rkey value of the first extension header (RDS_EXTHDR_RDMA) is
confined within 64K, then the first five bytes of the extension
headers looks like:

2, 0, 0, rkey1, rkey0

The two consecutive zeros is misinterpreted as the end of the
RDS_EXTHDR_RDMA extension, the code insists on having a
RDS_EXTHDR_NONE marker between the extension, and therefore, the first
bytes of the extension header will look like:

2, 0, 4, len3, len2, len1, len0, r/w

When this message is received, the extension headers are parsed. Since
the first "2" designates the RDS_EXTHDR_RDMA extension header, the
following four bytes will be used as the rkey.  The rkey is corrupted,
and the MR will NOT be revoked, despite being created with the
use-once semantics.

Further, if the four bytes {0, 4, len3, len2} matches an existing
rkey, this will accidentally be invalidated.

Re-factored part of the extension header insertion, removed the
unnecessary len variable when inserting (as the type designates the
size of the extension), removed the idiosyncratic RDS_EXTHDR_NONE
marker between extension headers, and most important, using the length
of the extension derived from its type during the parsing to find the
next spot for an extension header.

Fixes: "rds: new extension header: rdma bytes"
Signed-off-by: Håkon Bugge <haakon.bugge@oracle.com>
Signed-off-by: Somasundaram Krishnasamy <somasundaram.krishnasamy@oracle.com>
Signed-off-by: Allison Henderson <allison.henderson@oracle.com>